requirements.txt
gradio
huggingface_hub
pillow
python-dotenv
pydantic
requests
